feat: checker submits task (#19)
* feat: checker submits task

* refactor: simplify `common-raw` by generic

No more annoying `decode_*` or `encode_*`! Only `Serialize` and
`FromRaw` now.

* refactor: simplify `common` by generic

Prepare for moving pattern checking functions to `code-cell`.
Anyway, this is faster.

* refactor: move things from `common` to `code-cell`

Why?
It's not easy to maintain two library at the same time. Since only
`code-cell` handles main logic in our design, it's not necessary to
place these helpers in `common`.
May the cohesion be with you.